+++ Plugins/RoleMembers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # RoleMembers Changelog
 
+### 0.0.2
+
+ - Update for discord changes
+ - Add member count to popout
+ - Add role colors to context menu
+ - Add pseudo-role "@everyone" as an option
+
 ### 0.0.1
 
  - Initial release