+++ Plugins/ServerSearch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ServerSearch Changelog
 
+### 0.0.2
+
+ - Update for discord changes
+
 ### 0.0.1
 
  - Initial release